requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
gunicorn==21.2.0
daphne==4.1.2